backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
alembic>=1.14.0
psycopg2-binary>=2.9.10
redis>=5.2.0
httpx>=0.28.0
pydantic[email]>=2.10.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
stripe>=11.0.0
slowapi>=0.1.9
jinja2>=3.1.4
python-multipart>=0.0.17
